Remove cyclic dependency between public and private part of MozView
Currently 3 public view implementations (QuickMozView, QGraphicsView,
QOpenGLWebPage) use the same common private part,
QGraphicsMozViewPrivate. In all 3 cases the public view implementations
create an instance of QGrapicsMozViewPrivate in their constructors and
remove it in desctructor. At the same time a pointer to the public view
implementaton is storred in QGraphicsMozViewPrivate::q variable.
Unfortunately the QScopedPointer template class is used to manage the
pointer. As a result when public view is destroyed it calls delete on
it's private part, which in turn removes public view once again when
QScopedPointer goes out of scope. The public view implementation is
destroyed twice.

This patch replaces QScopedPointer with QPointer which should ensure
each public view type is only destroyed once.

Signed-off-by: Piotr Tworek <piotr.tworek@jollamobile.com>