[qtmozembed] Filter out orientations events with 500ms interval. Contributes to JB#35447
If multiple orientations are done quickly enough, we end up adding
unwanted work to the Gecko engine. This 500ms filter gives rather good
results when trying to quickly queue orientation change events. Problem
is that expensive reflows might take considerable amount of time and
we should not execute too many of them in row if not really needed.

In case orientation change event is cancelled that is informed
for the embedder.