[qtmozembed] Expose atXBeginning/atXEnd/atYBeginning/atYEnd properties. Contributes to JB#36023
This commit adds properties which expose whether the webview viewport
is panned to the LHS, RHS, top, or bottom.  Note that these are not
mutually exclusive panning states, depending on the size of the content
area.

These properties can be used to allow or disallow page stack navigation
or pulley-menu operation.

Contributes to JB#36023