[qtmozembed] Add javascriptEnabled property. Fixes JB#36018
Add a global setting to disable javascript execution.

Currently we don't have proper API to register to
listen preference changes. Thus, it is possible that
mJavascriptEnabled might be in an inconsistent state.

This can happen because there are multiple ways to change
preference value:

"about:config"
"prefs.js"
embedlite-components provides an API
qtmozembed API

To solve this I think that only viable option is
that we register preference observes for values
that QMozEngineSettings is interested of.