[qtmozembed] Expose SSL/TLS security info. Contributes to JB#40089
Provides a new inteface exposing the SSL/TLS and certificate information
in a form consumable by QML. Useful for providing security info to the
user. This presents the deserialised info provided by the gecko engine.

The certificate details are provided as a property map, allowing them to
be consumed by the CertificateDetailsPage of jolla-settings-system for
presentation to the user.

qmozview handles ownership and management of the QMozSecurity object.