Increase the margin of error the scroll boundary properties.
The x and y position after zooming in and out again often falls not
quite on zero by a margin greater the the scroll epsilon which meant
a scroll position which by all appearances was at a boundary reported
otherwise.

The ScrollOffset reported by gecko is an integer and will return exactly
0 in the same circumstances, but without a corresponding width and
height it's only good for the top and left edges.