[qtmozembed] Change QmlMozContext to singleton QMozContext. JB#54022 OMP#JOLLA-82
Only meaningful leftover from the previous QmlMozContext creatable
qml type is and was getenv function. As that is and was solely used by
test cases, the getenv is now available for test cases via singleton
TestHelper.

If we need test helper utilities in future, the TestHelper thingie
is a possible place to implement those helpers.