[qtmozembed] Support the "hidden" flag
Passes the "hidden" flag through from the gecko-dev EmbedLite API and
exposes it as a "hidden" flag to sailfish-browser (or whatever is
consuming the qtmozembed interface).

This is useful for printing, since the print routines need a window to
clone the page content into for the print. This chanage allows the tab
used to clone into to be hidden by the user-interface, which is
convenient because there's never any need for the user to interact with
it directly.