[qtmozembed] Don't skip pages without user interaction on goBack/Forward. JB#62664
The second parameter, added in ESR91, is aRequireUserInteraction.
Meaning that as true it skipped the pages that had no user interaction.

This broke a sailfish-browser unit tests for forward/backward navigation
and also made the back button act surprising in the Browser app when
it started jumping over pages in the history.