[qtmozembed] Fix tst_searchengine unit test. JB#62731
For some reason the test was testing that in addition for custom
search engine working, it will also do the search as lowercased.
With ESR91 the lowercasing doesn't happen, but for me that would feel
even better. Anyway let's not make here assumption on whether it's done
or not. If the search engine is triggered that's already good enough.