[qtmozembed] Avoid running gecko shutdown on all the unit tests. JB#62626
There is now some problem on the shutdown sequence that makes
qmlmoztestrunner segfault, making all the unit tests fail
even if the main case it was intended to test succeeded.

So making the full shutdown now opt-in and having it tested in the
basicview case. Gets the problem still visible, but not confusing it
with tests for other functionality.